pdf-service/requirements.txt
flask
pillow
numpy
requests